+++ src/Symfony/Bridge/Doctrine/CHANGELOG.md
@@ -10,3 +10,5 @@ CHANGELOG
  * DoctrineType now caches its choice lists in order to improve performance
  * DoctrineType now uses ManagerRegistry::getManagerForClass() if the option "em" is not set
  * UniqueEntity validation constraint now accepts a "repositoryMethod" option that will be used to check for uniqueness instead of the default "findBy"
+ * [BC BREAK] the DbalLogger::log() visibility has been changed from public to
+   protected